+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Unreleased
 ----------
 
+- Support signed JWT credentials. ([#14](https://github.com/persona-id/googleauth-extras/pull/14))
+
 - Drop support for Ruby 2.7 & 3.0, update test dependencies. ([#15](https://github.com/persona-id/googleauth-extras/pull/15))
 
 0.4.0